+++ tauro/blueprints/api_oauth2_v1/endpoints/consultar_turnos_unidad.py
@@ -37,18 +37,24 @@ class ConsultarTurnosUnidad(Resource):
 
         # Consultar los turnos...
         # - Filtrar por unidad,
-        # - Filtrar por los estados EN ESPERA y PASE A VENTANILLA,
+        # - Filtrar por los estados EN ESPERA o PASE A VENTANILLA o ATENDIENDO,
         # - Filtrar por el estatus A (activo),
         # - Y ordenar por el nombre de tipo de turno ATENCIÓN URGENTE, CON CITA, NORMAL y luego por el número del turno
         turnos = (
             Turno.query.join(TurnoEstado)
             .join(TurnoTipo)
             .filter(Turno.unidad_id == unidad.id)
-            .filter(TurnoEstado.nombre != "COMPLETADO", TurnoEstado.nombre != "CANCELADO")
+            .filter(
+                or_(
+                    TurnoEstado.nombre == "EN ESPERA",
+                    TurnoEstado.nombre == "PASE A VENTANILLA",
+                    TurnoEstado.nombre == "ATENDIENDO",
+                )
+            )
             .filter(Turno.estatus == "A")
             .order_by(
-                # 1. Prioridad por estado: PASE A VENTANILLA primero (valor 0), el resto después (valor 1)
-                case((TurnoEstado.nombre == "PASE A VENTANILLA", 0), else_=1),
+                # 1. Prioridad por estado: EN ESPERA primero (valor 0), el resto después (valor 1)
+                case((TurnoEstado.nombre == "EN ESPERA", 0), else_=1),
                 # 2. Dentro de cada grupo, ordenar por número de turno
                 Turno.numero,
             )
@@ -67,12 +73,17 @@ class ConsultarTurnosUnidad(Resource):
                 ),
             ).model_dump()
 
-        # Consultar Último turno en estado 'PASE A VENTANILLA'
+        # Consultar Último turno en estado 'ATENDIENDO' o 'ATENDIENDO EN CUBÍCULO'
         ultimo_turno_atendiendo = (
             Turno.query.join(TurnoEstado)
             .join(TurnoTipo)
             .filter(Turno.unidad_id == unidad.id)
-            .filter(or_(TurnoEstado.nombre == "PASE A VENTANILLA", TurnoEstado.nombre == "ATENDIENDO EN CUBICULO"))
+            .filter(
+                or_(
+                    TurnoEstado.nombre == "ATENDIENDO",
+                    TurnoEstado.nombre == "ATENDIENDO EN CUBICULO",
+                )
+            )
             .filter(Turno.estatus == "A")
             .order_by(TurnoTipo.nivel, Turno.numero)
             .first()